# src/main/resources/static/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI Documentation using Java Springboot","description":"OpenAPI Documentation with Swagger UI using Java Springboot","contact":{"name":"Kaushik Natua","email":"kaushik.natua@hyland.com"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"OpenAPI Documentation with Swagger UI using Java Springboot","description":"CRUD OpenAPIs in Java Springboot to CRUD Operations"},{"name":"Actuator","description":"Monitor and interact","externalDocs":{"description":"Spring Boot Actuator Web API Documentation","url":"https://docs.spring.io/spring-boot/docs/current/actuator-api/html/"}}],"paths":{"/api/accounts/update":{"put":{"tags":["OpenAPI Documentation with Swagger UI using Java Springboot"],"summary":"Update Account Details REST API","description":"REST API to update Customer &  Account details based on a account number","operationId":"updateAccountDetails","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"200":{"description":"HTTP Status OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}},"500":{"description":"HTTP Status Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}}}}},"/api/accounts/create":{"post":{"tags":["OpenAPI Documentation with Swagger UI using Java Springboot"],"summary":"Create Account REST API","description":"REST API to create new Customer &  Account informations","operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"201":{"description":"HTTP Status CREATED","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}}}}},"/actuator/loggers/{name}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'loggers-name'","operationId":"loggers-name","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}},"post":{"tags":["Actuator"],"summary":"Actuator web endpoint 'loggers-name'","operationId":"loggers-name_2","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","enum":["TRACE","DEBUG","INFO","WARN","ERROR","FATAL","OFF"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/accounts/fetch":{"get":{"tags":["OpenAPI Documentation with Swagger UI using Java Springboot"],"summary":"Fetch Account Details REST API","description":"REST API to fetch Customer &  Account details based on a mobile number","operationId":"fetchAccountDetails","parameters":[{"name":"mobileNumber","in":"query","required":true,"schema":{"pattern":"(^$|[0-9]{10})","type":"string"}}],"responses":{"200":{"description":"HTTP Status OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/actuator":{"get":{"tags":["Actuator"],"summary":"Actuator root web endpoint","operationId":"links","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}},"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}}}}}}},"/actuator/threaddump":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'threaddump'","operationId":"threaddump","responses":{"200":{"description":"OK","content":{"text/plain;charset=UTF-8":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/scheduledtasks":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'scheduledtasks'","operationId":"scheduledtasks","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/metrics":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'metrics'","operationId":"metrics","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/metrics/{requiredMetricName}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'metrics-requiredMetricName'","operationId":"metrics-requiredMetricName","parameters":[{"name":"requiredMetricName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/mappings":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'mappings'","operationId":"mappings","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/loggers":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'loggers'","operationId":"loggers","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/info":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'info'","operationId":"info","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/heapdump":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'heapdump'","operationId":"heapdump","responses":{"200":{"description":"OK","content":{"application/octet-stream":{"schema":{"type":"object"}}}}}}},"/actuator/health":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'health'","operationId":"health","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/health/**":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'health-path'","operationId":"health-path","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/env":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'env'","operationId":"env","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/env/{toMatch}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'env-toMatch'","operationId":"env-toMatch","parameters":[{"name":"toMatch","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/configprops":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'configprops'","operationId":"configprops","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/configprops/{prefix}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'configprops-prefix'","operationId":"configprops-prefix","parameters":[{"name":"prefix","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/conditions":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'conditions'","operationId":"conditions","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/caches":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'caches'","operationId":"caches","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Actuator"],"summary":"Actuator web endpoint 'caches'","operationId":"caches_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/actuator/caches/{cache}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'caches-cache'","operationId":"caches-cache","parameters":[{"name":"cache","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Actuator"],"summary":"Actuator web endpoint 'caches-cache'","operationId":"caches-cache_2","parameters":[{"name":"cache","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/actuator/beans":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'beans'","operationId":"beans","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/api/accounts/delete":{"delete":{"tags":["OpenAPI Documentation with Swagger UI using Java Springboot"],"summary":"Delete Account & Customer Details REST API","description":"REST API to delete Customer &  Account details based on a mobile number","operationId":"deleteAccountDetails","parameters":[{"name":"mobileNumber","in":"query","required":true,"schema":{"pattern":"(^$|[0-9]{10})","type":"string"}}],"responses":{"200":{"description":"HTTP Status OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}},"500":{"description":"HTTP Status Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"}}}}}}}},"components":{"schemas":{"Accounts":{"required":["accountNumber","accountType","branchAddress"],"type":"object","properties":{"accountNumber":{"type":"integer","description":"Account Number of Eazy Bank account","format":"int64"},"accountType":{"type":"string","description":"Account type of Eazy Bank account","example":"Savings"},"branchAddress":{"type":"string","description":"Eazy Bank branch address"}},"description":"Schema to hold Account information"},"Customer":{"required":["email","name"],"type":"object","properties":{"name":{"maxLength":30,"minLength":5,"type":"string","description":"Name of the customer","example":"Eazy Bytes"},"email":{"type":"string","description":"Email address of the customer","example":"tutor@eazybytes.com"},"mobileNumber":{"pattern":"(^$|[0-9]{10})","type":"string","description":"Mobile Number of the customer","example":"9345432123"},"account":{"$ref":"#/components/schemas/Accounts"}},"description":"Schema to hold Customer and Account information"},"Response":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"statusCode":{"type":"string","description":"Status code in the response","example":"200"},"statusMsg":{"type":"string","description":"Status message in the response","example":"Request processed successfully"}},"description":"Schema to hold successful response information"},"Link":{"type":"object","properties":{"href":{"type":"string"},"templated":{"type":"boolean"}}}}}}